pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "page analyzer"
authors = ["ErKir <kirillerohin@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
validators = "^0.20.0"


[tool.poetry.scripts]
page_analyzer = "page_analyzer.app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
